requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Computer Vision and OCR - HEADLESS version (no GUI dependencies)
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
PyPDF2==3.0.1

# AI/ML
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0